requirements.txt
asgiref==3.9.1
Django==5.2.4
django-dotenv==1.4.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
